+++ packages/azure/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v2.0.0] packages/azure
 ### Releases
 - [PyPI v2.0.0] https://pypi.org/project/mistralai-azure/2.0.0 - packages/azure
+
+## 2026-05-21 15:27:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.2 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.0] packages/azure
+### Releases
+- [PyPI v2.1.0] https://pypi.org/project/mistralai-azure/2.1.0 - packages/azure